Move code to a src dir.
+ create_project.py moved to src/main.py
+ Added root level create_project.py which calls src/main.main
+ Added src/__init__.py
+ Updated README
